requirements.txt
tensorflow
numpy
matplotlib
seaborn
pandas
streamlit
Pillow
PyYAML
scikit-learn
